Move Kale tags manipulation function to utils
